requirements.txt
anyascii==0.3.2
asgiref==3.8.1
autopep8==2.3.1
beautifulsoup4==4.12.3
bleach==4.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.1.1
django-allauth==64.2.1
django-cogwheels==0.3
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-dirtyfields==1.9.3
django-extensions==3.2.3
django-filter==24.3
django-flags==5.0.13
django-formtools==2.5.1
django-import-export==4.1.1
django-livereload-server==0.5.1
django-modeladmin-reorder==0.3.1
django-modelcluster==6.3
django-modeltranslation==0.19.9
django-multiselectfield==0.1.13
django-permissionedforms==0.1
django-silk==5.2.0
django-taggit==5.0.1
django-tinymce==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gprof2dot==2024.6.6
httplib2==0.22.0
idna==3.8
import-export==0.3.1
joblib==1.4.2
l18n==2021.3
laces==0.1.1
livereload==2.7.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
nltk==3.9.1
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pillow_heif==0.18.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydotplus==2.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
regex==2024.7.24
requests==2.32.3
rosetta==0.3
rsa==4.9
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
telepath==0.3.1
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wagtail==6.2.1
wagtail-cache==2.5.1
wagtail-color-panel==1.5.0
wagtail-flags==5.3.1
wagtail-instance-selector==3.0.1
wagtail-markdown==0.12.1
wagtail-modeltranslation==0.15.1
wagtail-seo==2.5.0
wagtail-sharing==2.12.1
wagtail-table-block==2.0.0
wagtail-transfer==0.9.4
wagtail_content_import==0.12.0
wagtailcodeblock==1.29.0.2
wagtailmedia==0.15.2
wagtailmenus==4.0.1
wagtailterms==0.1.3
webencodings==0.5.1
Werkzeug==3.0.4
Willow==1.8.0
